feat question fix

+++ server/src/main/java/pro/stackOverFlow/question/dto/QuestionDto.java
@@ -33,8 +33,7 @@ public class QuestionDto {
         private long questionId;
         private String title;
         private String content;
-
-        private int viewCount;
+        private long viewCount;
     }
